+++ .github/workflows/web-lint.yml
@@ -1,0 +1,42 @@
+name: 🧹 Web Lint
+
+on:
+  push:
+    paths:
+      - 'web/**'
+      - '.github/workflows/web-lint.yml'
+  pull_request:
+    paths:
+      - 'web/**'
+    types: [opened, synchronize, reopened]
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  lint:
+    name: 🧹 ESLint & Type Check
+    runs-on: ubuntu-latest
+    steps:
+      - name: 📥 リポジトリをチェックアウト
+        uses: actions/checkout@v4
+
+      - name: 🟢 Node.jsのセットアップ
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+          cache: 'npm'
+          cache-dependency-path: web/package-lock.json
+      
+      - name: 📦 依存関係をインストール
+        working-directory: web
+        run: npm ci
+
+      - name: 🧹 ESLintを実行
+        working-directory: web
+        run: npm run lint
+
+      - name: 🔍 TypeScript型チェック
+        working-directory: web
+        run: npm run build
